feat(demoPet): 添加数据统计功能并优化系统
- 新增数据统计相关接口和页面
- 优化订单列表和首页界面- 添加运营总览和累计数据统计
- 重构底部导航栏组件
- 优化数据展示样式

+++ demoPet/src/main/java/com/tgwei/demopet/demos/web/service/PetsInfoService.java
@@ -14,7 +14,7 @@ public class PetsInfoService {
     private PetsInfoMapper petsInfoMapper;
 
 
-    public List<PetsInfo> getAllPetsInfo(Integer ownerId) {
+    public List<PetsInfo> getAllPetsInfo(Long ownerId) {
         return petsInfoMapper.getAllPetsInfo(ownerId);
     }
